Color Coding & Jewel Recipes
Jewel recipes are started, Crafting recipes work when tested but are ENABLED=0 for now.

Color codes/naming schema are changed and working when tested on Runes, Chipped-Perfect Gems of all types.  Gold on the ground is abbreviated to a yellow "G" for less screen clutter as well.
Bolts, Arrows and Cracked items are black and Abbreviated (to hide as best as possible) but still clickable.
TP & ID Scrolls, Antidotes, Stamina potions, and Thawing potions are abbreviated and color coded.

No color coding on important affixes (of Teleportation, Golemlord's, Rose Branded) yet, but happy to add that if desired.